backend completed

+++ backend/controllers/product.controller.js
@@ -33,12 +33,16 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success:false, message:"Invalid Product ID"});
+    }
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product Deleted"});
     } catch (error) {
         console.log("Error in deleting product");
-        res.status(404).json({success:false, message: "Product not Found."});
+        res.status(500).json({success:false, message: "Server Error."});
     }
 }
